Kos Capacity in cfg
let the user change harddrive size in the part cfg

it needs this in the cfg MODULE
{
name = kOSProcessor
kOSCapacity = 10000
}

+++ Persistance/Harddisk.cs
@@ -6,9 +6,9 @@ namespace kOS.Persistance
 {
     public class Harddisk : Volume
     {
-        public Harddisk(int size)
+        public Harddisk(float size)
         {
-            Capacity = size;
+            Capacity = (int)size;
         }
 
         public Harddisk(ConfigNode node)
